+++ server/DataAccess/Models/JwtLastToken.cs
@@ -4,9 +4,9 @@ namespace server.DataAccess.Models;
 
 public class JwtLastToken
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
-    [ForeignKey("UserId")] public int UserId { get; set; }
+    [ForeignKey("UserId")] public long UserId { get; set; }
 
     public DateTime ExpiryDateTime { get; set; }
     public string Token { get; set; }
